必殺技ゲージの描画実装 レベル別に色変更、WindowSizeから描画サイズの設定可能

+++ X-MEN_Team_Project/Player.h
@@ -4,17 +4,17 @@
 
 namespace PlayerInitializeInfo
 {
-	const int MAX_HP = 100;
+	const float MAX_HP = 100;
 }
 
 class Player :
     public MelLib::GameObject
 {
 private:
-	int hp;
+	float hp;
 
 	// ïKéEãZÉQÅ[ÉW
-	int ultimateSkillValue;
+	float ultimateSkillValue;
 
 	bool isBarrier;
 	bool isThrowingBall;
